Updated Implementatieplan and Debugging
Removed some mistakes in Implementatieplan. As for the debugging, I still have the problem that the facial paramters are not shown correctly. It would also seem that the default implementation of the scaling step does nothing, so I did some minor editing. Now the default implementation actually scales the image, but it suffers from the same problems when showing the facial paramters. I think it has something to do whith the part of the program that shows the facial paramters: it seems it doen't to account for the scaling correctly.